Added vmLetExpr.go
